Added my MATLAB geometric multigrid implementation (mainly for debugging purposes), which does either:
1) Classical geometric multigrid, with rediscretizing every level and calling the (old, non-canonical) version of LaMEM at every level and reading in the stiffness matrixes.

2) Galerkin multigrid, which only reads in the finest grid stiffness matrix and forms coarser grids by Galerkin coarsening.

Restriction/Prolongation can be done matrix-based (using the same matrixes as used for coarsening) or through geometric interpolation (i.e. by calling interp3 in matlab).
Any LaMEM setup can be tested, but the default hardcoded one is a falling block one with a viscosity contrast.
We use jacobi smoothing during the V-cycle.

Note that this routine currently only works for the (old) version of LaMEM, which had dummy-layers. It should be modified to work with canonical. 

The main (starting) routines are:
* Stokes_LaMEM_Classical_GMG
* Stokes_LaMEM_Galerkin_GMG
 

git-svn-id: svn+ssh://gaia.geo.uni-mainz.de/local/home/lkausb/svn/LaMEM/trunk@5488 26174531-3d5a-4d31-aba0-7a6658d032cb