started to integrate the GMT-API as an optional output format. I didn't update the makefile yet, because I rather want to start a new branch as soon as the repository has moved to GIT.
It seems that the integration is straightforward. This will enable us to (1) perform post-processing (filtering, computing correlations with tomographies, using all the inbuilt functionality of GMT) on the fly and (2) to create nice GMT graphics at the end of a simulation.

We should aim to use the enhanced file formats such as (parallel) necdf and hdf5.


The documentation is available here:
http://gmt.soest.hawaii.edu/doc/5.1.0/GMT_API.html

git-svn-id: svn+ssh://gaia.geo.uni-mainz.de/local/home/lkausb/svn/LaMEM/trunk@5872 26174531-3d5a-4d31-aba0-7a6658d032cb