made changes the manner in which random noise is introduced to MATLAB generated particles.
1) We now generate random noise in the mesh setup routine:

[X,Y,Z,xcoor,ycoor,zcoor, Xpart, Ypart, Zpart] = FDSTAGMeshGeneratorMatlab(npartx,nparty,npartz,fname, RandomNoise)

where noise is added if RandomNoise = logical(1)
The coordinates of all particles are in given in the 3D array: 

	[Xpart, Ypart, Zpart]

which should be used to set Phases, temperatures etc.

2) The particle coordinates should be added to the A structure as:
	A.Xpart  =  Xpart;
	A.Ypart  =  Ypart;
	A.Zpart  =  Zpart;

where Xpart etc. might need to be permuted to fit the size of the A.Phase array (in my code setup is Xpart = permute(Xpart,[2 1 3]);)




git-svn-id: svn+ssh://gaia.geo.uni-mainz.de/local/home/lkausb/svn/LaMEM/trunk@6020 26174531-3d5a-4d31-aba0-7a6658d032cb