Recovered functionality of the default block preconditioner.
The performance for the multi-sinker problem is terrible compared to the
coupled solver. With more than one order of magnitude viscosity contrast
it basically stops converging, unless direct solver for the viscosity
block is used. Even in this case it’s very slow. Coupled solver easily
handles three orders.

I have an expectation that whatever advantage BFBT will bring for this
problem, it will be no match to the performance of the coupled solver.

Options can be found in BuildInSetups/FallingSpheres_Multigrid.dat