*changed the way the recently introduced pressure shift (pShift) is propagated through the code.
This is important in cases with a free slip upper BC, but with phase transitions that have a Clapeyron slope and are thus dependent on the pressure. With free slip, P is only defined up to a constant, and as a result the pressure @ the upper boundary is typically not zero. One can add a pressure shift to (somewhat) compensate for this effect, even when there is no guarantee that the required shift in P (to have average zero P @ the upper boundary) remains the same throughout a simulation.