Short commit: 1) Previously i was looping element by element and reducing it. Now I'm trying to integrate a collective operation but I got a nasty message: Assertion failed in file src/mpi/coll/helper_fns.c at line 84: FALSE memcpy argument memory ranges overlap, dst_=0x7ffd1654ac60 src_=0x7ffd1654ac10 len_=20000 I tried to create only sequential vectors: 2) The problem appears at the beginning. So, whenever is solved, we can simply copy and paste the stuff to the other reduction operation.
The alghoritm is simple:
Create a set of vectors that are known by all the processor (seq? MPI? Scatter?)
Each vector initially has the coordinates of the passive tracers.
During the advection it changes the position, the temperature, pressure (eventually Phase) of the marker that has the coordinate within the current processor
The passive tracer outside the processor boundary are set to have a value equal to the negative to the doublemax.
The reduction operation find the maximum value

Apperently this create a memoryoverlap:
my guess is that I'm creating the wrong object, but I do not still have found an alternative object that can do what I want